+++ src/Mapbender/CoreBundle/Element/Type/ControlButtonAdminType.php
@@ -6,12 +6,21 @@ namespace Mapbender\CoreBundle\Element\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Validator\Constraints\NotBlank;
 
 class ControlButtonAdminType extends AbstractType
 {
     public function getParent()
     {
         return 'Mapbender\CoreBundle\Element\Type\BaseButtonAdminType';
+    }
+
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults(array(
+            'application' => null,
+        ));
     }
 
     /**
@@ -25,7 +34,8 @@ class ControlButtonAdminType extends AbstractType
             ))
             ->add('target', 'Mapbender\CoreBundle\Element\Type\TargetElementType', array(
                 'application' => $options['application'],
-                'required' => false,
+                'required' => true,
+                'constraints' => array(new NotBlank()),
             ))
         ;
     }