+++ src/Mapbender/ManagerBundle/Component/ElementFormFactory.php
@@ -10,7 +10,9 @@ use Mapbender\CoreBundle\Entity\Element;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 use Symfony\Component\Form\FormFactoryInterface;
+use Symfony\Component\Form\FormRegistryInterface;
 use Symfony\Component\Form\FormTypeInterface;
+use Symfony\Component\Validator\Constraints\NotBlank;
 
 /**
  * Service for element configuration and acl forms.
@@ -25,21 +27,27 @@ class ElementFormFactory extends BaseElementFactory
     protected $container;
     /** @var bool */
     protected $strict;
+    /** @var FormRegistryInterface */
+    protected $formRegistry;
 
     /**
      * @param FormFactoryInterface $formFactory
      * @param ElementInventoryService $inventoryService
      * @param ContainerInterface $container
+     * @param FormRegistryInterface $formRegistry
      * @param bool $strict
      */
     public function __construct(FormFactoryInterface $formFactory,
                                 ElementInventoryService $inventoryService,
-                                ContainerInterface $container, $strict = false)
+                                ContainerInterface $container,
+                                FormRegistryInterface $formRegistry,
+                                $strict = false)
     {
         parent::__construct($inventoryService);
         $this->formFactory = $formFactory;
         $this->container = $container;
         $this->setStrict($strict);
+        $this->formRegistry = $formRegistry;
     }
 
     public function setStrict($enable)
@@ -56,17 +64,28 @@ class ElementFormFactory extends BaseElementFactory
     {
         // Create base form shared by all elements
         $formType = $this->formFactory->createBuilder('Symfony\Component\Form\Extension\Core\Type\FormType', $element, $options);
-        $formType
-            ->add('title', 'Symfony\Component\Form\Extension\Core\Type\TextType')
-        ;
         $this->migrateElementConfiguration($element);
+        $titleConstraints = array(
+            new NotBlank(),
+        );
+        $formType
+            ->add('title', 'Mapbender\ManagerBundle\Form\Type\ElementTitleType', array(
+                'constraints' => $titleConstraints,
+            ))
+        ;
         $configurationType = $this->getConfigurationFormType($element);
 
-        $options = array('application' => $element->getApplication());
+        $options = array();
 
         $componentClassName = $this->getComponentClass($element);
         $twigTemplate = $componentClassName::getFormTemplate();
         $options['label'] = false;
+
+        $resolvedType = $this->formRegistry->getType($configurationType);
+        if ($resolvedType->getOptionsResolver()->isDefined('application')) {
+            // Only pass the "application" option if the form type requires / declares it.
+            $options['application'] = $element->getApplication();
+        }
 
         $formType->add('configuration', $configurationType, $options);